==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
# jupyter magic command
# makes jupyter reload everytime we write a new import statement
%reload_ext autoreload
%autoreload 2

# Updates the sys.path so we can import modules from the src folder
# Solution obtained from https://stackoverflow.com/questions/9427037/relative-path-not-working-even-with-init-py

import os, sys
sys.path.insert(0, os.path.abspath(".."))

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
