==> notebooks/06_baseline_model.ipynb <==
In this notebook we define three different rule-based baseline models.

If our machine learning model has a final test mean absolute error that is larger than the results of these models, then we are not gaining any additional value by adding the complexity of running a machine learning model.

# Updates the sys.path so we can import modules from the src folder
# Solution obtained from https://stackoverflow.com/questions/9427037/relative-path-not-working-even-with-init-py

import os, sys
sys.path.insert(0, os.path.abspath(".."))

import pandas as pd 
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import numpy as np

# defines a baseline model
# this model predicts the next hour demand to be the same as in the previous hour
class BaselineModelPreviousHour:
    """
    Prediction = actual demand observed in the last hour
    """
    # there is no fit function since this is a rule-based model, not a machine learning model
    # it does not require training 
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test[f'rides_previous_1_hour']

# runs the baseline model to get predictions on test data
model = BaselineModelPreviousHour()
predictions = model.predict(X_test)
predictions

from sklearn.metrics import mean_absolute_error

# evaluates the baseline model on test data
# the performance metric for this model will be the mean absolute error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

import pandas as pd
import numpy as np

# defines a new model that accounts for weekly seasonality
# it predicts the next hour demand to be the same as the last week's demand at that same hour.
class BaselineModelPreviousWeek:
    """
    Prediction = actual demand observed at t - 7 days
    """
    # there is no fit function since this is a rule-based model, not a machine learning model
    # it does not require training 
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test[f'rides_previous_{7*24}_hour']

# runs the second baseline model on test data to get predictions
model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)

# evaluates the second baseline model performance on test data
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# defines a model that predicts the next hour demand as an average of different time frames.
# estimates the average for the previous 7, 14, 21 and 28 days

class BaselineModelLast4Weeks:
    """
    Prediction = actual demand observed at t - 7 days, t - 14 days, t - 21 days, t - 28 days
    """

    # there is no fit function since this is a rule-based model, not a machine learning model
    # it does not require training 
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        """"""
        return 0.25*(
            X_test[f'rides_previous_{7*24}_hour'] + \
            X_test[f'rides_previous_{2*7*24}_hour'] + \
            X_test[f'rides_previous_{3*7*24}_hour'] + \
            X_test[f'rides_previous_{4*7*24}_hour']
        )

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
