==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
### This notebook creates a scikit-learn feature engineering and lightgbm fit pipeline
#### Based on: https://towardsdatascience.com/how-to-quickly-design-advanced-sklearn-pipelines-3cc97b59ce16

# Updates the sys.path so we can import modules from the src folder
# Solution obtained from https://stackoverflow.com/questions/9427037/relative-path-not-working-even-with-init-py

import os, sys
sys.path.insert(0, os.path.abspath(".."))

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

# splits the data into train and test sets
from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X
     


# wraps the average rides function into a scikit-learn data transformation step
from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)
     

# applies the .fit() and .transform() methods to the original dataframe
# using our average rides_function to add the new features
add_feature_average_rides_last_4_weeks.fit_transform(X_train)


# extracts day of the week, and the hour of the day, from the pickup_hour column

from sklearn.base import BaseEstimator, TransformerMixin

# creates our own class that inherits from Baseestimator and TransformerMixin
class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    # implements the .fit() method, however there are no parameters to fit in this case
    def fit(self, X, y=None):
        return self
    
    #implements the .transform() method
    def transform(self, X, y=None):
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_['pickup_hour'].dt.hour
        X_["day_of_week"] = X_['pickup_hour'].dt.dayofweek
        
        return X_.drop(columns=['pickup_hour'])

# calls the .fit_transform() method on the training data
add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

import lightgbm as lgb

# puts together the three steps into a scikit-learn pipeline
from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)
     

# compute predictions and evaluates errors
predictions = pipeline.predict(X_test)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
